feat:登录成功后，从本地数据库中获取好友列表，在聊天列表中显示

+++ model/friend.hpp
@@ -10,7 +10,7 @@ class Friend : public User
 {
 public:
     explicit Friend() = default;
-    explicit Friend(int wxid, QString name, QString image, bool mute, QString time, int count, QList<Message> messages):
+    explicit Friend(uint64_t wxid, QString name, QString image, bool mute, QString time, int count, QList<Message> messages):
         User(wxid, name.toStdString(), image.toStdString()), m_mute(mute), m_time(time), m_count(count), m_messages(messages){}
